change rm

+++ window_cmd/rm.py
@@ -4,4 +4,7 @@ import os
 import glob
 all_file=glob.glob(sys.argv[1])
 for item in all_file:
-    os.remove(item)
+    if os.path.isfile(item):
+        os.remove(item)
+    if os.path.isdir(item):
+        os.removedirs(item)
